three-tier-web/requirements.txt
ansicolors==1.1.8
appnope==0.1.0
attrs==20.2.0
aws-cdk.assets==1.64.1
aws-cdk.aws-applicationautoscaling==1.64.1
aws-cdk.aws-autoscaling-common==1.64.1
aws-cdk.aws-certificatemanager==1.64.1
aws-cdk.aws-cloudwatch==1.64.1
aws-cdk.aws-codeguruprofiler==1.64.1
aws-cdk.aws-ec2==1.64.1
aws-cdk.aws-efs==1.64.1
aws-cdk.aws-elasticloadbalancingv2==1.64.1
aws-cdk.aws-events==1.64.1
aws-cdk.aws-iam==1.64.1
aws-cdk.aws-kms==1.64.1
aws-cdk.aws-lambda==1.64.1
aws-cdk.aws-logs==1.64.1
aws-cdk.aws-rds==1.64.1
aws-cdk.aws-route53==1.64.1
aws-cdk.aws-s3==1.64.1
aws-cdk.aws-s3-assets==1.64.1
aws-cdk.aws-sam==1.64.1
aws-cdk.aws-secretsmanager==1.64.1
aws-cdk.aws-sqs==1.64.1
aws-cdk.aws-ssm==1.64.1
aws-cdk.cloud-assembly-schema==1.64.1
aws-cdk.core==1.64.1
aws-cdk.cx-api==1.64.1
aws-cdk.region-info==1.64.1
backcall==0.1.0
bleach==3.1.4
boto3==1.14.19
botocore==1.17.19
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.0
constructs==3.0.4
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
git-remote-codecommit==1.15.1
gitdb==4.0.5
GitPython==3.1.3
idna==2.10
importlib-metadata==1.6.0
ipykernel==5.3.3
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isengardcli==0.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsii==1.12.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==1.0.14
ptyprocess==0.6.0
publication==0.0.3
Pygments==2.6.1
PyInquirer==1.0.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
regex==2020.6.8
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.4
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
-e .
